refactor: moved css to tailwind and moved data to separate file
