--- src/components/sections/Faqs.css
@@ -1,13 +1,0 @@
-.collapse {
-  @apply rounded-lg border border-gray-300 bg-white;
-}
-
-.collapse-title {
-  @apply cursor-pointer text-xl font-medium;
-}
-
-.collapse,
-.collapse-title,
-.collapse-content * {
-  @apply text-black;
-}
